+++ src/viur/datastore/__init__.py
@@ -2,11 +2,12 @@ from viur.datastore.types import KEY_SPECIAL_PROPERTY, DATASTORE_BASE_TYPES, Sor
 	QueryDefinition, Key, currentDbAccessLog
 from viur.datastore.query import Query
 from viur.datastore.utils import fixUnindexableProperties, normalizeKey, keyHelper, IsInTransaction, GetOrInsert,\
-	encodeKey, acquireTransactionSuccessMarker
+	encodeKey, acquireTransactionSuccessMarker, startDataAccessLog, endDataAccessLog
 from viur.datastore.transport import Get, Put, Delete, RunInTransaction, Collision, AllocateIDs
 from viur.datastore.config import conf as config
 
 __all__ = ["KEY_SPECIAL_PROPERTY", "DATASTORE_BASE_TYPES", "SortOrder", "SkelListRef", "Entity", "QueryDefinition",
 		   "Key", "Query", "fixUnindexableProperties", "normalizeKey", "keyHelper", "Get", "Put", "Delete",
 		   "RunInTransaction", "IsInTransaction", "currentDbAccessLog", "GetOrInsert", "encodeKey",
-		   "acquireTransactionSuccessMarker", "Collision", "AllocateIDs", "config"]
+		   "acquireTransactionSuccessMarker", "Collision", "AllocateIDs", "config", "startDataAccessLog",
+		   "endDataAccessLog"]