+++ src/viur/datastore/query.py
@@ -78,7 +78,9 @@ class Query(object):
 		self._fulltextQueryString: Union[None, str] = None
 		self.lastCursor = None
 		if not kind.startswith("viur") and not kwargs.get("_excludeFromAccessLog"):
-			currentDbAccessLog.get(set()).add(kind)
+			accessLog = currentDbAccessLog.get()
+			if isinstance(accessLog, set):
+				accessLog.add(kind)
 
 	def setFilterHook(self, hook: Callable) -> Optional[Callable]:
 		"""