+++ tests/__init__.py
@@ -2,4 +2,5 @@ from .basic import BasicFunctionTest
 from .transactions import TransactionTest
 from .queryvalues import QueryValuesTest
 from .querycustomfunctions import QueryCustomFunctionsTest
+from .dataaccesslog import DataAccessLogTest